+++ lib/ecto/adapter/transactions.ex
@@ -6,8 +6,20 @@ defmodule Ecto.Adapter.Transactions  do
   use Behaviour
 
   @doc """
-  Runs the function on the given repo inside a transaction.
+  Runs the given function inside a transaction. Returns `{ :ok, value }` if the
+  transaction was successful where `value` is the value return by the function
+  or `{ :error, value }` if the transaction was rolled back where `value` is the
+  value given to `rollback/1`.
+
   See `Ecto.Repo.transaction/1`.
   """
-  defcallback transaction(Ecto.Repo.t, fun) :: any
+  defcallback transaction(Ecto.Repo.t, fun) :: { :ok, any } | { :error, any }
+
+  @doc """
+  Rolls back the current transaction. The transaction will return the value
+  given as `{ :error, value }`.
+
+  See `Ecto.Repo.rollback/1`.
+  """
+  defcallback rollback(Ecto.Repo.t, any) :: no_return
 end