+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * Add Decimal support
   * Add 'distinct' query expression
   * Add `Validator.bin_dict/2`
+  * Add `Ecto.Repo.rollback` for explicit transaction rollback
 
 * Bug fixes
   * Fix association functions resetting the entity when manually loading associated entities